Added error handling

+++ .build/CleanSourcesFolder.ps1
@@ -1,4 +1,13 @@
 task Clean_Sources_Folder {
-    Remove-Item -Path $SourcePath\DSCResources -Recurse -Force
-    Remove-Item -Path $SourcePath\DSCResources.yml -Force
+
+    if (Test-Path -Path $SourcePath\DSCResources)
+    {
+        Remove-Item -Path $SourcePath\DSCResources -Recurse -Force
+    }
+
+    if (Test-Path -Path $SourcePath\DSCResources.yml)
+    {
+        Remove-Item -Path $SourcePath\DSCResources.yml -Force -ErrorAction SilentlyContinue
+    }
+
 }
